pyproject.toml
[project]
name = "vp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Tony Francis", email = "tony@dream3d.com" }
]
requires-python = ">=3.12"
dependencies = [
    "ansible>=11.3.0",
    "ansible-runner>=2.4.0",
    "fabric>=3.2.2",
    "invoke>=2.2.0",
    "loguru>=0.7.3",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "tenacity>=9.0.0",
    "typer>=0.15.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipython>=9.0.2",
    "ruff>=0.9.10",
]
